end of chp 4

+++ src/process.cpp
@@ -1,15 +1,32 @@
 #include "libsdb/error.hpp"
+#include "libsdb/pipe.hpp"
+#include <cerrno>
 #include <csignal>
+#include <cstdlib>
+#include <cstring>
 #include <libsdb/process.hpp>
 #include <memory>
+#include <string>
 #include <sys/ptrace.h>
 #include <sys/types.h>
 #include <sys/wait.h>
 #include <unistd.h>
 
-std::unique_ptr<sdb::process> sdb::process::launch(std::filesystem::path path)
+namespace
+{
+void exit_with_errno(sdb::pipe &channel, std::string const &prefix)
+{
+    auto message = prefix + ": " + std::strerror(errno);
+
+    channel.write(reinterpret_cast<std::byte *>(message.data()), message.size());
+    exit(-1);
+}
+} // namespace
+
+std::unique_ptr<sdb::process> sdb::process::launch(std::filesystem::path path, bool debug)
 {
     pid_t pid;
+    pipe channel(true);
 
     if ((pid = fork()) < 0)
     {
@@ -18,23 +35,39 @@ std::unique_ptr<sdb::process> sdb::process::launch(std::filesystem::path path)
 
     if (pid == 0)
     {
+        channel.close_read();
+
         // Inside child process.
 #if defined(__APPLE__) && defined(__MACH__)
-        if (ptrace(PT_TRACE_ME, 0, nullptr, 0) < 0)
+        if (debug and ptrace(PT_TRACE_ME, 0, nullptr, 0) < 0)
 #elif defined(__linux__)
-        if (ptrace(PTRACE_TRACEME, 0, nullptr, nullptr) < 0)
+        if (debug and ptrace(PTRACE_TRACEME, 0, nullptr, nullptr) < 0)
 #endif
         {
-            error::send_errno("Tracing failed");
+            exit_with_errno(channel, "Tracing failed");
         }
         if (execlp(path.c_str(), path.c_str(), nullptr) < 0)
         {
-            error::send_errno("execlp failed");
+            exit_with_errno(channel, "execlp failed");
         }
     }
 
-    std::unique_ptr<sdb::process> proc(new process(pid, true));
-    proc->wait_on_signal();
+    channel.close_write();
+    auto data = channel.read();
+    channel.close_read();
+
+    if (data.size() > 0)
+    {
+        // wait for the child process to terminate and issue an error.
+        waitpid(pid, nullptr, 0);
+        auto chars = reinterpret_cast<char *>(data.data());
+        error::send(std::string(chars, chars + data.size()));
+    }
+
+    std::unique_ptr<sdb::process> proc(new process(pid, true, debug));
+
+    if (debug)
+        proc->wait_on_signal();
 
     return proc;
 }
@@ -55,7 +88,7 @@ std::unique_ptr<sdb::process> sdb::process::attach(pid_t pid)
         error::send_errno("Could not attach");
     }
 
-    std::unique_ptr<sdb::process> proc(new process(pid, false));
+    std::unique_ptr<sdb::process> proc(new process(pid, false, true));
     proc->wait_on_signal();
 
     return proc;
@@ -99,22 +132,24 @@ sdb::process::~process()
     {
         int status;
 
-        // In order for PTRACE_DETACH to work, the inferior process must've
-        // been stopped.
-        if (state_ == process_state::running)
+        if (is_attached_)
         {
-            kill(pid_, SIGSTOP);
-            waitpid(pid_, &status, 0);
-        }
+            // In order for PTRACE_DETACH to work, the inferior process must've
+            // been stopped.
+            if (state_ == process_state::running)
+            {
+                kill(pid_, SIGSTOP);
+                waitpid(pid_, &status, 0);
+            }
 
 #if defined(__APPLE__) && defined(__MACH__)
-        ptrace(PT_DETACH, pid_, nullptr, 0);
+            ptrace(PT_DETACH, pid_, nullptr, 0);
 #elif defined(__linux__)
-        ptrace(PTRACE_DETACH, pid_, nullptr, nullptr);
+            ptrace(PTRACE_DETACH, pid_, nullptr, nullptr);
 #endif
-        // Once de-attached, let the process continue running.
-        kill(pid_, SIGCONT);
-
+            // Once de-attached, let the process continue running.
+            kill(pid_, SIGCONT);
+        }
         if (terminate_on_end_)
         {
             kill(pid_, SIGKILL);
